+++ client/src/components/account/options/Favorites.vue
@@ -2,7 +2,7 @@
   <div id = "section">
     <h2>MY FAVORITES</h2>
     <div id = "favorites" v-if="!isLoading">
-      <EventCard v-for = "(event, index) in events" :key = "index" :event = "event"></EventCard>
+      <MovieCard v-for = "(event, index) in events" :key = "index" :event = "event"></MovieCard>
     </div>
   </div>
 </template>
@@ -13,7 +13,7 @@ import MovieCard from "@/components/events/Movies/MovieCard.vue";
 
 export default {
   name: "MyFavorites",
-  components: {EventCard: MovieCard},
+  components: {MovieCard},
   data() {
     return {
       events: {},
@@ -52,14 +52,25 @@ export default {
 #favorites {
   display: flex;
   flex-direction: row;
+  justify-content: center;
   gap: 1.3rem;
   flex-wrap: wrap;
   padding: 0;
 }
 
+#card {
+  width: 17%;
+  border: 0.1rem solid #DDD;
+}
+
+#card:hover {
+  box-shadow: 0.01rem 0.01rem 0.2rem 0.05rem rgba(0, 0, 0, 0.05);;
+}
+
 h2 {
   font-size: 1.6rem;
   margin: 2rem 0.1rem 1rem 0;
+  text-align: center;
 }
 
 </style>