+++ client/src/components/events/Movies/ViewTicket.vue
@@ -1,21 +1,22 @@
 <template>
   <div v-if = "!isLoading" class = "root_component">
-    <HomeHeader></HomeHeader>
-    <div class="ticket">
-      <GenerateTicket :ticket = "ticket"></GenerateTicket>
+    <div class = "fill" ref = "ticket">
+      <div class = "ticket">
+        <GenerateTicket :ticket = "ticket" @save = "save"></GenerateTicket>
+      </div>
     </div>
+    <button @click = "() => {this.$router.push('/')}">BACK TO HOME</button>
   </div>
 </template>
 
 <script>
-
-import HomeHeader from "@/components/home/HomeHeader.vue";
 import GenerateTicket from "@/components/events/Movies/GenerateTicket.vue";
 import axios from "axios";
+import html2pdf from "html2pdf.js";
 
 export default {
   name: "ViewTicket",
-  components: {GenerateTicket, HomeHeader},
+  components: {GenerateTicket},
   data() {
     return {
       ticket: {},
@@ -25,36 +26,30 @@ export default {
   created() {
     this.isLoading = true;
     axios.post(`${process.env.VUE_APP_SERVER}/api/account/tickets/movies/${this.$route.params.ticketId}`, {}).then(res => {
-          this.ticket = res.data.ticket;
-          this.isLoading = false;
-        }).catch(err => {
+      this.ticket = res.data.ticket;
+      this.isLoading = false;
+    }).catch(err => {
       alert(err.response.data.message);
     })
+  },
+  methods: {
+    save() {
+      const ticket = this.$refs.ticket;
+      ticket.classList.add("pdf-mode");
+
+      html2pdf().set({
+        margin: 0.5,
+        filename: `swifticket-${this.ticket.ticketId}.pdf`,
+        image: {type: 'jpeg', quality: 0.98},
+        html2canvas: {scale: 2, useCORS: true},
+        jsPDF: {unit: 'in', format: 'letter', orientation: 'portrait'},
+      }).from(ticket).save().finally(() => {
+        ticket.classList.remove("pdf-mode");
+      });
+    }
   }
 }
 </script>
 
-<style scoped>
-
-.root_component {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-.ticket {
-  width: 33%;
-  height: auto;
-  border-radius: 1rem;
-  padding: 0.3rem;
-  box-shadow: -0.05rem 0.05rem 0.8rem 0 rgba(0, 0, 0, 0.15);
-  opacity: 1;
-  transition: opacity 1s ease;
-  background-color: rgb(233, 106, 106);
-  transform: scale(1.05) translateY(3%);
-}
-
-</style>
+<style scoped src = "../../../styles/button.css"></style>
+<style scoped src = "../../../styles/ticket.css"></style>